requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiwolf==1.2.0
anytree==2.8.0
appnope==0.1.2
async-timeout==4.0.2
attrs==22.2.0
b==1.0.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.1.0
cli-helpers==2.2.1
click==8.1.3
configobj==5.0.6
cryptography==37.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.4
entrypoints==0.3
filelock==3.6.0
fonttools==4.28.3
frozenlist==1.3.3
idna==3.4
importlib-resources==5.7.1
ipykernel==6.6.0
ipython==7.30.1
ita==0.2.10
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
load-dotenv==0.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
multidict==6.0.4
mycli==1.25.0
nest-asyncio==1.5.4
numpy==1.21.4
openai==0.27.2
packaging==21.3
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2022.4.30
platformdirs==2.5.2
prompt-toolkit==3.0.24
ptyprocess==0.7.0
PuLP==2.6.0
pyaes==1.6.1
pycparser==2.21
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==3.0.6
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
pyzmq==22.3.0
requests==2.28.2
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
urllib3==1.26.15
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
yarl==1.8.2
